BMP: fix RLE 4 decoding
Previously when absolute mode specified 4n+1 pixels it was being mishandled.
In this case `i` should be incremented twice at the end but it was not
being incremented at all.  This case is fairly rare in the wild as three
nibbles are being wasted on padding.

Modify the compression of the test case to have a case of absolute mode
specifying 5 pixels.